+++ .eslintrc.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-const { escapeRegExp } = require( 'lodash' );
 const glob = require( 'glob' ).sync;
 const { join } = require( 'path' );
 
@@ -17,7 +16,8 @@ const { version } = require( './package' );
  * @type {string}
  */
 const majorMinorRegExp =
-	escapeRegExp( version.replace( /\.\d+$/, '' ) ) + '(\\.\\d+)?';
+	version.replace( /\.\d+$/, '' ).replace( /[\\^$.*+?()[\]{}|]/g, '\\$&' ) +
+	'(\\.\\d+)?';
 
 /**
  * The list of patterns matching files used only for development purposes.
@@ -83,6 +83,7 @@ module.exports = {
 							'capitalize',
 							'chunk',
 							'clamp',
+							'cloneDeep',
 							'compact',
 							'concat',
 							'countBy',
@@ -94,14 +95,18 @@ module.exports = {
 							'differenceWith',
 							'dropRight',
 							'each',
+							'escapeRegExp',
 							'extend',
 							'findIndex',
 							'findKey',
 							'findLast',
+							'first',
 							'flatMap',
 							'flatten',
 							'flattenDeep',
+							'forEach',
 							'fromPairs',
+							'has',
 							'identity',
 							'invoke',
 							'isArray',
@@ -118,6 +123,7 @@ module.exports = {
 							'isUndefined',
 							'keyBy',
 							'keys',
+							'last',
 							'lowerCase',
 							'mapKeys',
 							'maxBy',
@@ -127,6 +133,7 @@ module.exports = {
 							'nth',
 							'once',
 							'overEvery',
+							'partial',
 							'partialRight',
 							'random',
 							'reject',
@@ -134,6 +141,7 @@ module.exports = {
 							'reverse',
 							'size',
 							'snakeCase',
+							'sortBy',
 							'startCase',
 							'startsWith',
 							'stubFalse',
@@ -145,6 +153,7 @@ module.exports = {
 							'toString',
 							'trim',
 							'truncate',
+							'unionBy',
 							'uniq',
 							'uniqBy',
 							'uniqueId',