+++ packages/block-library/src/navigation/edit/overlay-menu-icon.js
@@ -2,8 +2,17 @@
  * WordPress dependencies
  */
 import { SVG, Rect } from '@wordpress/primitives';
+import { Icon, menu, moreVertical, moreHorizontal } from '@wordpress/icons';
 
-export default function OverlayMenuIcon() {
+export default function OverlayMenuIcon( { icon } ) {
+	if ( icon === 'menu' ) {
+		return <Icon icon={ menu } />;
+	} else if ( icon === 'more-vertical' ) {
+		return <Icon icon={ moreVertical } />;
+	} else if ( icon === 'more-horizontal' ) {
+		return <Icon icon={ moreHorizontal } />;
+	}
+
 	return (
 		<SVG
 			xmlns="http://www.w3.org/2000/svg"