+++ packages/block-library/src/post-template/edit.js
@@ -17,7 +17,7 @@ import {
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { Spinner } from '@wordpress/components';
-import { store as coreStore } from '@wordpress/core-data';
+import { store as coreStore, useEntityRecords } from '@wordpress/core-data';
 
 const TEMPLATE = [
 	[ 'core/post-title' ],
@@ -85,6 +85,13 @@ export default function PostTemplateEdit( {
 			inherit,
 			taxQuery,
 			parents,
+			pages,
+			// We gather extra query args to pass to the REST API call.
+			// This way extenders of Query Loop can add their own query args,
+			// and have accurate previews in the editor.
+			// Noting though that these args should either be supported by the
+			// REST API or be handled by custom REST filters like `rest_{$this->post_type}_query`.
+			...restQueryArgs
 		} = {},
 		queryContext = [ { page: 1 } ],
 		templateSlug,
@@ -94,6 +101,18 @@ export default function PostTemplateEdit( {
 } ) {
 	const [ { page } ] = queryContext;
 	const [ activeBlockContextId, setActiveBlockContextId ] = useState();
+
+	let categorySlug = null;
+	if ( templateSlug?.startsWith( 'category-' ) ) {
+		categorySlug = templateSlug.replace( 'category-', '' );
+	}
+	const { records: categories, hasResolved: hasResolvedCategories } =
+		useEntityRecords( 'taxonomy', 'category', {
+			context: 'view',
+			per_page: -1,
+			_fields: [ 'id' ],
+			slug: categorySlug,
+		} );
 
 	const { posts, blocks } = useSelect(
 		( select ) => {
@@ -155,13 +174,21 @@ export default function PostTemplateEdit( {
 				if ( templateSlug?.startsWith( 'archive-' ) ) {
 					query.postType = templateSlug.replace( 'archive-', '' );
 					postType = query.postType;
+				} else if ( !! categorySlug && hasResolvedCategories ) {
+					query.taxQuery = {
+						category: categories.map( ( { id } ) => id ),
+					};
+					taxQuery = query.taxQuery;
 				}
 			}
 			// When we preview Query Loop blocks we should prefer the current
 			// block's postType, which is passed through block context.
 			const usedPostType = previewPostType || postType;
 			return {
-				posts: getEntityRecords( 'postType', usedPostType, query ),
+				posts: getEntityRecords( 'postType', usedPostType, {
+					...query,
+					...restQueryArgs,
+				} ),
 				blocks: getBlocks( clientId ),
 			};
 		},
@@ -181,7 +208,11 @@ export default function PostTemplateEdit( {
 			templateSlug,
 			taxQuery,
 			parents,
+			restQueryArgs,
 			previewPostType,
+			categories,
+			categorySlug,
+			hasResolvedCategories,
 		]
 	);
 	const blockContexts = useMemo(